requirements.txt
fastapi==0.115.5
aiogram==3.15.0
flake8==7.1.1
asyncpg==0.30.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
uvicorn==0.32.1
Jinja2==3.1.4
python-multipart==0.0.18
itsdangerous==2.2.0